Updated tests, added a priority queue
Updated the tests to reflect the functionality added in verion 0.4.0, and then
updated the library to actually pass all those tests. Notably, there was a
problem with loading and dumping to and from files, but it's been resolved.

This version also sees the introduction of a priority queue, with unique elements.
While it is potentially problematic to be tied to the uniqueness constraing,
one can always inject arbitrary ids for repeated values. Perhaps not the most
efficient method in the world, but it exposes one more list type in redis.